requirements.txt
reportlab>=3.5.60
pycairo>=1.20.0,<1.21
